test: split the key-event suite to satisfy SwiftLint
KeyEventTests had grown to a 355-line class body, five lines past
SwiftLint's 350-line error threshold, which was the one serious violation
failing the lint job. Split it along the seam it already had: routing and
dispatch stay in KeyEventTests, focus-ring membership and traversal move
to FocusRingTests, and the shared fixtures move to a KeyEventTestCase base
class so neither file re-declares them.

No test bodies changed. swiftlint now exits clean; the remaining 21
violations are pre-existing warnings, none serious.
